Checkpoint Updates | COMPONENT Rewrite to Functional Component! | DO NOT PUBLISH
I'm still in the middle of the process if I should keep those another GraphQL queries (not fragments!) seperated or inlined from respective page components. As I kept getting error that it was not a page component such as Index.JS

I think using creatingPages API in gatsby-node.js should solve the problem.